refactor(epub): keep the CSS module's stated policy honest, and tidy the seams (#251)
The `css` module doc claimed it implements no box model and honours three
properties. Both stopped being true — it honours nine, and vertical margins are
the box model. It now says what it does: no cascade, no layout, and exactly one
box-model property, with the reason that one is worth it and the others are not.

`Length` gets a hand-written `PartialEq` so its `Eq` is sound. `parse_length`
rejects non-finite values, but the variants are public and a `NaN` built outside
would have broken reflexivity for every `Eq` derived on top of it.

`is_zero_length` and `parse_length` were scanning a numeric prefix character for
character in two spellings ten lines apart; both now share `split_number`, which
also stops them allocating a `String` they immediately discard. The four
property arms in `BlockStyle::apply` that used `if let Some(..)` now read
`x = parse(..).or(x)` like the five added beside them — one idiom for one policy.

Also: `merge_cell_flows` dropped a `column_x` computation nothing could read, and
its epsilon and doc now state the real reason it exists and what it does not
promise. `content.rs` moves its tests to `content_tests.rs`, the pattern the
crate's other modules already follow, bringing it from 1223 lines to 655.
`Length` and `PageBreak` join the re-exports, since `BlockStyle`'s public fields
are typed with them. Repairs a doc comment for `LayoutOpts` that #194 had cut in
half, and trims the reuse argument that had been made three times to the one
place the invariant lives.
